+++ origin-token-transfer/client/public/favicon.ico
