+++ origin-token-transfer/client/src/components/NavBar.js
