+++ origin-token-transfer/server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "origin-token-transfer",
+  "name": "origin-token-transfer-server",
   "version": "1.0.0",
   "description": "Tools for distributing Origin tokens",
   "engines": {
@@ -7,9 +7,8 @@
   },
   "main": "app/app.js",
   "scripts": {
-    "client": "cd src/client && SKIP_PREFLIGHT_CHECK=true npm start",
-    "server": "cd src/server && nodemon server.js",
-    "test": "mocha -r @babel/register -r @babel/polyfill -t 10000 --exit test"
+    "start:dev": "nodemon src/app.js",
+    "test": "mocha -r @babel/register -r @babel/polyfill -t 10000"
   },
   "author": "",
   "license": "ISC",
