+++ origin-token-transfer/server/src/scripts/.sequelizerc
