+++ origin-token-transfer/server/src/scripts/show_grants.sh
