+++ origin-token-transfer/server/src/scripts/wipe_and_initialize_db.sh
